refactor: move `IntradayRouter` & `APIRouter` into production.
